stop using an arraylist as storage for the clients
Instead store _pointers_ to clients now.

Using an ArrayList as storage for the Client struct makes the
bookkeeping difficult because we work a pointers to clients and removing
a client from the storage means invalidating pointers to existing
clients.

An ArrayList of Client _pointers_ individually allocated allows us to
have Client pointers which never change and still have a way to iterate
over all clients currently connected.